--- config copy.toml
@@ -1,55 +1,0 @@
-# Load config file and parameters
-filename = 'input\Alex_45 Day_Comments_ref - Copy.xlsx'
-sheetname = 'Comments Database'
-
-[columns]
-sort = [
-  "FSOR section heading level 1",
-  "FSOR section heading level 2",
-  "FSOR section heading level 3",
-]
-
-# Numbered sort columns if sort-sort is set to 'worksheet'
-numbered_sort = [
-  "value1",
-  "value2",
-  "value3",
-]
-
-[columns.commentresponse]
-comment = 'CommentText'
-response = 'draft Agency Response'
-
-[columns.other]
-tag = 'File Name'
-
-[sort]
-# Specify sorting of lowest-level (where comments and responses are).
-# Comment Count (descending): 'count', Alphabetical (ascending): anything else
-type = ''
-
-# //TODO
-# Where to get sort data from. 'worksheet' for columns specified under numbered sort,
-# or 'config' for sort data from this configuration file.
-sort = 'config'
-
-# Specify custom sort of level 1. Use empty list [] for no custom sort.
-# For headings not listed here, it will be at the end and sorted as specified above.
-custom_level_1 = [
-  "Support"
-]
-# This is a lump-all of all subheadings past Level 1. This really shines as a way to 
-# bring something to the first position of all subheadings - like a "General" category
-# or something. But, it can also be used to 
-custom_level_2 = [
-  "Air compressors"
-]
-
-[other]
-# Specify whether what row header begins on. Assumes data starts immediately after.
-header_row = 1
-# Specify starting heading level (1 means begin at Heading 1)
-outline_level = 2
-# Specify whether to remove leading and trailing spaces at the end of paragraphs.
-# Also removes double-spaces.
-clean = true
