+++ log.log
@@ -1,7 +1,7 @@
-2022-06-22 10:12:43.094 [INFO] Logging initialized.
-2022-06-22 10:12:43.094 [INFO] Reading sheet...
-2022-06-22 10:12:44.077 [INFO] Creating comment-response document...
-2022-06-22 10:12:44.609 [INFO] Comment-response document created: output\CommentResponseSection.docx
-2022-06-22 10:12:44.609 [INFO] Creating automark document...
-2022-06-22 10:12:44.651 [INFO] Automark document created: output\AutoMark.docx
-2022-06-22 10:12:44.651 [INFO] Finished.
+2022-06-22 13:49:02.166 [INFO] Logging initialized.
+2022-06-22 13:49:02.166 [INFO] Reading sheet...
+2022-06-22 13:49:03.159 [INFO] Creating comment-response document...
+2022-06-22 13:49:03.705 [INFO] Comment-response document created: output\CommentResponseSection.docx
+2022-06-22 13:49:03.705 [INFO] Creating automark document...
+2022-06-22 13:49:03.750 [INFO] Automark document created: output\AutoMark.docx
+2022-06-22 13:49:03.750 [INFO] Finished.
