--- ooxml/__init__.py
@@ -1,7 +1,0 @@
-from .spreadsheet import *
-from .sharedstrings import *
-
-__all__ = [
-    "SpreadSheetML",
-    "RichText"
-]
